pyproject.toml
[project]
name = "langchain-harness"
version = "0.2.0"
description = "Runtime-configurable deep-agent harness on LangChain + Anthropic ADK (Claude Opus 4.7)"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.1.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "anthropic>=0.40.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.8.0",
]

[project.optional-dependencies]
tracing = [
    "langsmith>=0.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]

[project.scripts]
langchain-harness = "langchain_harness.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/langchain_harness"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
